login personnalisé

+++ app/Http/Controllers/Admin/UsersController.php
@@ -6,6 +6,7 @@ use App\Role;
 use App\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 
 class UsersController extends Controller
 {
@@ -62,6 +63,11 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
+       if( Gate::denies('edit-users'))
+       {
+           return redirect()->route('admin.users.index');
+       }
+
         $roles = Role::all();
         return view('admin.users.edit',[
             'user'=>$user,
@@ -82,6 +88,10 @@ class UsersController extends Controller
         // dd($request);
         $user->roles()->sync($request->roles);
 
+        $user->name= $request->nom;
+        $user->email = $request->email; 
+        $user->save();
+
         return redirect()->route('admin.users.index');
     }
 
@@ -93,6 +103,11 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        if( Gate::denies('delete-users'))
+        {
+            return redirect()->route('admin.users.index');
+        }
+
         $user->roles()->detach();
         $user->delete();
         return redirect(route('admin.users.index'));
